+++ programs/zk-token-proof-tests/tests/process_transaction.rs
@@ -18,7 +18,7 @@ use {
 const VERIFY_INSTRUCTION_TYPES: [ProofInstruction; 6] = [
     ProofInstruction::VerifyZeroBalance,
     ProofInstruction::VerifyWithdraw,
-    ProofInstruction::VerifyWithdrawWithheldTokens,
+    ProofInstruction::VerifyCiphertextCiphertextEquality,
     ProofInstruction::VerifyTransfer,
     ProofInstruction::VerifyTransferWithFee,
     ProofInstruction::VerifyPubkeyValidity,
@@ -61,17 +61,17 @@ async fn test_zero_balance() {
 }
 
 #[tokio::test]
-async fn test_withdraw_withheld_tokens() {
-    let elgamal_keypair = ElGamalKeypair::new_rand();
+async fn test_ciphertext_ciphertext_equality() {
+    let source_keypair = ElGamalKeypair::new_rand();
     let destination_keypair = ElGamalKeypair::new_rand();
 
     let amount: u64 = 0;
-    let withdraw_withheld_authority_ciphertext = elgamal_keypair.public.encrypt(amount);
+    let source_ciphertext = source_keypair.public.encrypt(amount);
 
-    let success_proof_data = WithdrawWithheldTokensData::new(
-        &elgamal_keypair,
+    let success_proof_data = CiphertextCiphertextEqualityProofData::new(
+        &source_keypair,
         &destination_keypair.public,
-        &withdraw_withheld_authority_ciphertext,
+        &source_ciphertext,
         amount,
     )
     .unwrap();
@@ -80,32 +80,32 @@ async fn test_withdraw_withheld_tokens() {
         public: ElGamalKeypair::new_rand().public,
         secret: ElGamalKeypair::new_rand().secret,
     };
-    let fail_proof_data = WithdrawWithheldTokensData::new(
+    let fail_proof_data = CiphertextCiphertextEqualityProofData::new(
         &incorrect_keypair,
         &destination_keypair.public,
-        &withdraw_withheld_authority_ciphertext,
+        &source_ciphertext,
         amount,
     )
     .unwrap();
 
     test_verify_proof_without_context(
-        ProofInstruction::VerifyWithdrawWithheldTokens,
+        ProofInstruction::VerifyCiphertextCiphertextEquality,
         &success_proof_data,
         &fail_proof_data,
     )
     .await;
 
     test_verify_proof_with_context(
-        ProofInstruction::VerifyWithdrawWithheldTokens,
-        size_of::<ProofContextState<WithdrawWithheldTokensProofContext>>(),
+        ProofInstruction::VerifyCiphertextCiphertextEquality,
+        size_of::<ProofContextState<CiphertextCiphertextEqualityProofContext>>(),
         &success_proof_data,
         &fail_proof_data,
     )
     .await;
 
     test_close_context_state(
-        ProofInstruction::VerifyWithdrawWithheldTokens,
-        size_of::<ProofContextState<WithdrawWithheldTokensProofContext>>(),
+        ProofInstruction::VerifyCiphertextCiphertextEquality,
+        size_of::<ProofContextState<CiphertextCiphertextEqualityProofContext>>(),
         &success_proof_data,
     )
     .await;
